added cart-quantity-input, update handling of quantity-input, update styles
