+++ layout/theme.liquid
@@ -19,7 +19,6 @@
       {{ content_for_layout }}
       <!-- SCRIPTS -->
       <script src="{{ 'utility-methods.js' | asset_url }}" defer="defer"></script>
-      <script src="{{ 'quantity-input.js' | asset_url }}" type="module"></script>
     </main>
     {% section 'footer' %}
   </body>
